requirements.txt
Django==1.11
djangorestframework==3.3.3
Pillow==4.3.0
